backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
asyncpg==0.31.0
beautifulsoup4==4.14.3
certifi==2026.2.25
click==8.3.2
et_xmlfile==2.0.0
fastapi==0.135.3
feedparser==6.0.12
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.43.0
xlrd==2.0.2
redis
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
